combo

+++ FYS_105/GameObject.pde
@@ -27,8 +27,8 @@ public class GameObject {
   float pillarPosX, pillarPosY, pillarWidth, pillarHeight;
 
   PVector spawnerPos0, spawnerPos1, spawnerPos2, spawnerPos3;
-
-  int scoreGain;
+  float scoreGain;
+  
 
   GameObject() 
   {
